pyproject.toml
[project]
name = "ai-mentor"
version = "0.1.0"
description = ""
authors = [
    {name = "dhanaabhirajk",email = "81916146+dhanaabhirajk@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "llama-index (>=0.12.11,<0.13.0)",
    "chainlit (>=2.0.4,<3.0.0)",
    "llama-index-vector-stores-lancedb (>=0.3.0,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
